Move createTextFragmentArg to its own file
